I am basicly sleepcoding

+++ FuelStation_Final/GasStation/GasStation/TransactionLines.cs
@@ -28,11 +28,8 @@ namespace FuelStation.Win
     public partial class frm_TransactionLines : Form
     {
         private readonly HttpClient _client;
-        // private List<TransactionListDTO> _transactionList = new();
         private List<TransactionLineListDTO> _translineList = new();
         private List<ItemListDTO> _itemList = new();
-        //private List<CustomerListDTO> _customerList = new();
-        //private List<EmployeeListDTO> _employeeList = new();
         private readonly TransactionHandler _transHandler = new();
         private TransactionLineListDTO _newLine = new();
         private TransactionListDTO _parentTransaction = new();
@@ -70,35 +67,37 @@ namespace FuelStation.Win
             col_ItemID.DataSource = _itemList.ToList();
             col_ItemID.ValueMember = "ID";
             col_ItemID.DisplayMember = "Description";
-
         }
 
         private async Task OnSave(TransactionLineListDTO newLine)
         {
+           // newLine = (TransactionLineListDTO)bsTransLine.Current;
             HttpResponseMessage response = null;
-
-            MessageBox.Show("Save cAlled");
-            // pushare ola ta transactionLines ston controller
-            //kane update to total sto transaction
+           // _translineList.Add(newLine);
+            response = await _client.PostAsJsonAsync("transactionLine", newLine);
+            if (response.IsSuccessStatusCode)
+            { MessageBox.Show("Save successful"); }
+            else
+            {
+              MessageBox.Show("Save unsuccessful");
+            }
 
         }
         //========================== Transaction Lines Buttons ============================
         #region add Button
         private void btn_trl_Add_Click(object sender, EventArgs e)
         {
-            //maybe a post before creating a new line?
-
             bsTransLine.AddNew();
             _newLine = bsTransLine.Current as TransactionLineListDTO;
             _newLine.TransactionId = _parentTransaction.ID;
-
+            
         }
         #endregion
 
         #region bussiness logic
         private void bsTransLine_CurrentItemChanged(object sender, EventArgs e)
         {
-            List<ItemListDTO> fuelItems = _itemList.Where(type => type.ItemType == ItemType.Fuel).ToList();  
+            List<ItemListDTO> fuelItems = _itemList.Where(type => type.ItemType == ItemType.Fuel).ToList();
             //==========================Auto-Fill Item Price=================================== 
 
 
@@ -118,11 +117,11 @@ namespace FuelStation.Win
 
             if (_newLine.NetValue > 20 && _translineList.Any(x => fuelItems.Any(item => item.ID == x.ItemID)))
             {
-                _newLine.DiscountPercent = 0.1M;
+                _newLine.DiscountPercent = 10;
             }
             else _newLine.DiscountPercent = 0;
 
-            _newLine.DiscountValue = _newLine.NetValue * _newLine.DiscountPercent;
+            _newLine.DiscountValue = (_newLine.NetValue * _newLine.DiscountPercent)/10;
             _newLine.TotalValue = _newLine.NetValue - _newLine.DiscountValue;
 
             //=============================End Gridview Edit=======================================   
@@ -134,99 +133,64 @@ namespace FuelStation.Win
 
         //END TRANSACTION BUTTON -- will decide what to send other than total value
         private void button1_Click(object sender, EventArgs e)
-        {
-            this.Hide();
-            Transactions_frm transact = new Transactions_frm(_newLine.TotalValue);
+        { decimal tmpTotal = 0;
+            this.Hide();//might need to send a list 
+           var totals = _translineList.Where(trl => trl.TotalValue != 0).Select(trl => trl.TotalValue).ToList();
+           foreach(var total in totals) 
+            { tmpTotal += total; } 
+           
+            Transactions_frm transact = new Transactions_frm(tmpTotal);
             transact.ShowDialog();
             this.Dispose();
         }
-
-
         private async void btn_trl_delete_Click(object sender, EventArgs e)
         {
             TransactionLineListDTO transLineDelete = (TransactionLineListDTO)bsTransLine.Current;
             HttpResponseMessage response = null;
-
-            response = await _client.DeleteAsync($"transaction/{transLineDelete.ID}");
-            _translineList.Remove(transLineDelete);
+            if (transLineDelete != null || transLineDelete.ID != 0)
+            {
+                response = await _client.DeleteAsync($"transaction/{transLineDelete.ID}");
+                _translineList.Remove(transLineDelete);
+                bsTransLine.RemoveCurrent();
+            }
             await SetControlProperties();
-
         }
-
         private void grv_TransactionLine_DataError(object sender, DataGridViewDataErrorEventArgs e)
         {
             e.Cancel = true;
         }
 
-        private void btn_trl_save_Click(object sender, EventArgs e)
+        private async void btn_trl_save_Click(object sender, EventArgs e)
         {
-            if (ValidateTransLines(_translineList, _itemList)){
-                OnSave(_newLine);
+            if (ValidateTransLines(_translineList, _itemList))
+            {
+               await OnSave(_newLine);
             }
-            else { 
-                MessageBox.Show("There is some error with overlaping fuel type Transaction lines"); 
+            else
+            {
+                MessageBox.Show("There is some error with overlaping fuel type Transaction lines");
             }
         }
-       
-        
         public bool ValidateTransLines(List<TransactionLineListDTO> list, List<ItemListDTO> itemList)
         {
             int counter = 0;
             List<ItemListDTO> fuelItems = itemList.Where(type => type.ItemType == ItemType.Fuel).ToList();
-            foreach(var tr in list)
+            foreach (var tr in list)
             {
-                foreach(var item in fuelItems)
+                foreach (var item in fuelItems)
                 {
-                    if(item.ID == tr.ItemID)
+                    if (item.ID == tr.ItemID)
                     {
                         counter++;
                     }
                 }
             }
-            if (counter>1){ return false; } else { return true; }
+            if (counter > 1) { return false; } else { return true; }
         }
 
-        /* private void grv_TransactionLine_CellValueChanged(object sender, DataGridViewCellEventArgs e)
-         {
-             if (e.ColumnIndex >= 0 && e.RowIndex >= 0 && grv_TransactionLine.Columns[e.ColumnIndex].Name == "ItemID")
-             {
-                 var currentItem = (TransactionLineListDTO)bsTransLine.Current;
-                 var fuelItems = _itemList.Where(i => i.ItemType == ItemType.Fuel).ToList();
-                 if (fuelItems.Any(fi => fi.ID == currentItem.ItemID))
-                 {
-                     MessageBox.Show("You cannot select an item of type fuel.", "Error");
-                     bsTransLine.ResetCurrentItem();
-                 }
-             }
-         }*/
+        private void frm_TransactionLines_Load(object sender, EventArgs e)
+        {
 
-        /* private void grv_TransactionLine_RowValidated(object sender, DataGridViewCellEventArgs e)
-         {
-             if (e.RowIndex >= 0)
-             {
-                 var currentItem = (TransactionLineListDTO)bsTransLine.List[e.RowIndex];
-                 var fuelItems = _itemList.Where(i => i.ItemType == ItemType.Fuel).ToList();
-                 if (fuelItems.Any(fi => fi.ID == currentItem.ItemID))
-                 {
-                     MessageBox.Show("You cannot select an item of type fuel.", "Error");
-                    // e.Cancel = true;
-                 }
-             }
-         }*/
-
-        /* private void grv_TransactionLine_RowValidating(object sender, DataGridViewCellCancelEventArgs e)
-         {
-
-             if (e.RowIndex >= 0)
-             {
-                 var currentItem = (TransactionLineListDTO)bsTransLine.List[e.RowIndex];
-                 var fuelItems = _itemList.Where(i => i.ItemType == ItemType.Fuel).ToList();
-                 if (fuelItems.Any(fi => fi.ID == currentItem.ItemID))
-                 {
-                     MessageBox.Show("You cannot select an item of type fuel.", "Error");
-                      e.Cancel = true;
-                 }
-             }
-         }*/
+        }
     }
 }
